Wrong StandIn filename when using a sequence pattern with motion blur enabled (Ticket #864)

+++ plugins/mtoa/nodes/shape/ArnoldStandIns.h
@@ -25,7 +25,7 @@ public:
    MString filename;
    MString geomLoaded;
    int mode;
-   float frame;
+   int frame;
    float frameOffset;
    bool useFrameExtension;
    bool useSubFrame;
